git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@16444 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ cell_based/src/mesh/PottsMeshWriter.cpp
@@ -207,7 +207,8 @@ void PottsMeshWriter<SPACE_DIM>::WriteFiles()
     // Write the element header
     unsigned num_elements = this->GetNumElements();
 
-    unsigned first_elem_attribute_value = (*(mpIters->pElemIter))->GetAttribute();
+    /// \todo  #2146 - this looks like a bug in the making - what if the attribute IS zero!?!?
+    double first_elem_attribute_value = (*(mpIters->pElemIter))->GetAttribute();
     if (first_elem_attribute_value != 0)
     {
         num_attr = 1;
